requirements.txt
Pillow
tqdm
numpy
